fix: reorganises project structure
